requirements.txt
fastapi
pylint
redis
requests
types-redis
ChemSpiPy
sgd-rest
uvicorn

